:hammer: refactor: app 이름 변경 기존: game 변경: reception
