+++ config/settings.py
@@ -23,7 +23,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'channels',
-    'game',
+    'reception',
     'corsheaders'
 ]
