+++ config/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path('api/game/', include('game.urls')),
+    path('api/reception/', include('reception.urls')),
 ]
